# openapi.json
{
  "components": {
    "schemas": {
      "CommandName": {
        "description": "Batchalign command name (e.g. `\"morphotag\"`, `\"align\"`).",
        "type": "string"
      },
...
      "DurationSeconds": {
        "description": "Duration measured in fractional seconds.",
        "format": "double",
        "type": "number"
      },
      "ErrorResponse": {
        "description": "Standard error response body, matching FastAPI's `HTTPException` format.\n\nAll non-2xx responses from the server use this shape so that clients\ncan parse errors uniformly. The one exception is\n[`ServerError::JobConflict`](crate::error::ServerError::JobConflict),\nwhich nests a `conflicts` array inside `detail`.",
        "properties": {
          "detail": {
            "description": "Human-readable description of what went wrong. For validation\nerrors this includes the specific field or constraint that failed.",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "JobId": {
        "description": "Server-assigned UUID (v4) for a job.",
        "type": "string"
      },
...
      "JobListItem": {
        "description": "Summary for job listing (`GET /jobs` response element).\n\nA lighter-weight projection of [`JobInfo`] without per-file statuses,\nsuitable for listing many jobs at once.",
        "properties": {
          "active_lease": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LeaseRecord",
                "description": "Active lease information when this job is currently claimed by a node."
              }
            ]
          },
          "command": {
            "$ref": "#/components/schemas/CommandName",
            "description": "Batchalign command (e.g. \"morphotag\", \"align\")."
          },
          "completed_at": {
            "description": "ISO 8601 timestamp of when the job reached a terminal state.",
            "type": [
              "string",
              "null"
            ]
          },
          "completed_files": {
            "description": "Number of files that finished successfully (`Done`).",
            "format": "int64",
            "type": "integer"
          },
          "duration_s": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DurationSeconds",
                "description": "Wall-clock duration in seconds.  `None` while the job is active."
              }
            ]
          },
          "error_files": {
            "description": "Number of files that ended in `Error` status.",
            "format": "int64",
            "type": "integer"
          },
          "job_id": {
            "$ref": "#/components/schemas/JobId",
            "description": "Server-assigned UUID (v4) for this job."
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageCode3",
            "description": "3-letter ISO language code.  Defaults to `\"eng\"` when absent in JSON."
          },
          "next_eligible_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnixTimestamp",
                "description": "When the job is deferred before execution, the earliest unix timestamp\nat which it should be retried. `None` when immediately eligible."
              }
            ]
          },
          "num_workers": {
            "description": "Number of concurrent Python workers used for this job.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "source_dir": {
            "description": "Client's original input directory path, for display.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle state of the job."
          },
          "submitted_at": {
            "description": "ISO 8601 timestamp of submission.",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by": {
            "description": "IP address or Tailscale identifier of the submitting client.",
            "type": [
              "string",
              "null"
            ]
          },
          "submitted_by_name": {
            "description": "Human-readable hostname of the submitting machine.",
            "type": [
              "string",
              "null"
            ]
          },
          "total_files": {
            "description": "Total number of files in this job.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "job_id",
          "status",
          "command",
          "total_files",
          "completed_files"
        ],
        "type": "object"
      },
...
      "JobSubmission": {
        "description": "`POST /jobs` request body.",
        "properties": {
          "before_paths": {
            "description": "Absolute paths to \"before\" files for incremental processing\n(paths_mode only). When non-empty, the diff engine compares each\nbefore file against its corresponding source_path and only\nreprocesses changed utterances.\n\nMust be the same length as `source_paths` when non-empty.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "command": {
            "$ref": "#/components/schemas/CommandName",
            "description": "Batchalign command (align, morphotag, etc.)."
          },
          "debug_traces": {
            "description": "When true, the server collects detailed algorithm traces for\nvisualization (DP alignment matrices, ASR pipeline stages, FA\ntimelines, retokenization mappings). Defaults to false — zero\noverhead when off.",
            "type": "boolean"
          },
          "display_names": {
            "description": "Human-readable filenames for display (paths_mode only, optional).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "files": {
            "description": "CHAT files to process.",
            "items": {
              "$ref": "#/components/schemas/FilePayload"
            },
            "type": "array"
          },
          "lang": {
            "$ref": "#/components/schemas/LanguageCode3",
            "description": "3-letter ISO language code."
          },
          "media_files": {
            "description": "Media filenames for the server to resolve from media_roots (transcribe only).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "media_mapping": {
            "description": "Key into server's media_mappings config (e.g. \"childes-data\").",
            "type": "string"
          },
          "media_subdir": {
            "description": "Subdirectory under the mapped root (e.g. \"Eng-NA/MacWhinney/0young-ASR\").",
            "type": "string"
          },
          "num_speakers": {
            "$ref": "#/components/schemas/NumSpeakers",
            "description": "Number of speakers."
          },
          "options": {
            "description": "Typed command options (engine selections, processing flags, etc.)."
          },
          "output_paths": {
            "description": "Absolute paths to write output files to (paths_mode only).",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "paths_mode": {
            "description": "When true, server reads/writes files directly via source_paths/output_paths.",
            "type": "boolean"
          },
          "source_dir": {
            "description": "Client's input directory path (for dashboard display).",
            "type": "string"
          },
          "source_paths": {
            "description": "Absolute paths to read input files from (paths_mode only).",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "command",
          "options"
        ],
        "type": "object"
      },
      "LanguageCode3": {
        "description": "3-letter ISO 639-3 language code (e.g. `\"eng\"`, `\"spa\"`).",
        "type": "string"
      },
...
      "StatusMessageResponse": {
        "description": "Response body for mutating operations that return a status confirmation\n(e.g. cancel, delete, restart).\n\nReturned with HTTP 200 on success. Clients can branch on `status` for\nprogrammatic handling and display `message` to users.",
        "properties": {
          "message": {
            "description": "Human-readable sentence describing the action taken (e.g.\n`\"Job abc123 has been cancelled.\"`).",
            "type": "string"
          },
          "status": {
            "description": "Machine-readable outcome token (e.g. `\"cancelled\"`, `\"deleted\"`,\n`\"restarted\"`). Suitable for switch/match in client code.",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "batchalign3 application: types, server, worker pool, cache",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "batchalign-app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "description": "Used by the dashboard and CLI `jobs` command. The response is intentionally\ncompact -- per-file detail is omitted and available via `GET /jobs/{id}`.",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobListItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List all jobs"
          }
        },
        "summary": "Return a summary of every job the server knows about (active, completed,\nfailed, or cancelled).",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Validates the command against built-in tasks and worker-advertised capabilities,\nstages input files (content mode) or records source paths (paths mode), detects\n`(submitted_by, filename)` conflicts with active jobs, and spawns a background\nrunner task that acquires workers and dispatches files. The response echoes the\njob back as `JobInfo` and includes the correlation ID in `X-Request-Id`.",
        "operationId": "submit_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Submitted job"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Accept a new processing job and begin execution.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "description": "Returns 409 if the job is still running -- the caller must cancel it first.\nDeleting a job removes it from the in-memory store and SQLite, and broadcasts\na `JobDeleted` event to connected WebSocket/SSE clients.",
        "operationId": "delete_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Job deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job still running"
          }
        },
        "summary": "Permanently remove a terminal job and its associated state.",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "This is the primary polling endpoint for the CLI: the client calls it\nrepeatedly until `status` reaches a terminal value.",
        "operationId": "get_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job detail"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Return full detail for a single job, including per-file statuses.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "description": "Fires the job's `CancellationToken`, which the runner checks between files.\nAlready-terminal jobs (completed, failed, cancelled) are no-ops. Interrupted\njobs can still be cancelled -- they are not considered terminal here because\nthey may be holding worker resources.",
        "operationId": "cancel_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusMessageResponse"
                }
              }
            },
            "description": "Cancel request accepted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Request cancellation of a running or queued job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/restart": {
      "post": {
        "description": "Only jobs in a terminal non-cancelled state can be restarted. The store\nresets per-file statuses and clears errors, then a fresh runner task is\nspawned. This is the primary recovery path after transient worker crashes\nor OOM kills.",
        "operationId": "restart_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobInfo"
                }
              }
            },
            "description": "Job restarted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not restartable"
          }
        },
        "summary": "Reset a failed or interrupted job back to `Queued` and re-run it.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results": {
      "get": {
        "description": "Blocks until the job has completed, failed, or been interrupted -- returns\n409 for jobs that are still running or queued. In content mode the response\nincludes the full CHAT text of each output file; in paths mode only filenames\nare returned (the files were written directly to disk by the runner).\nCancelled jobs are intentionally excluded because they may have incomplete\nresults.",
        "operationId": "get_results",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultResponse"
                }
              }
            },
            "description": "Results for terminal jobs"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "Retrieve all output files for a terminal job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/results/{filename}": {
      "get": {
        "description": "Waits for the individual file to reach a terminal status, not the whole job,\nso clients can fetch results incrementally as files finish. Supports\nstem-based matching (e.g., requesting `sample.wav` will find `sample.cha`)\nto handle the common case where the output extension differs from the input.",
        "operationId": "get_single_result",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Requested output filename",
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResult"
                }
              }
            },
            "description": "One file result"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found"
          }
        },
        "summary": "Retrieve the output for a single file within a job.",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/stream": {
      "get": {
        "description": "Sends:\n- `snapshot` event with current file statuses on connect\n- `file_update` events as files are processed\n- `complete` event when the job reaches a terminal state, then closes",
        "operationId": "stream_job",
        "parameters": [
          {
            "description": "Job identifier",
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Job not found"
          }
        },
        "summary": "SSE stream for real-time per-file progress of a single job.",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}